@@ -25,11 +25,11 @@ import java.util.Iterator;
 import java.util.List;
 
 import javax.xml.XMLConstants;
-import javax.xml.bind.Binder;
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.Marshaller;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.Binder;
+import jakarta.xml.bind.JAXBContext;
+import jakarta.xml.bind.Marshaller;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.validation.Schema;
@@ -180,7 +180,7 @@ public class EnvironmentContext extends ContextObject{
         super(ENVIRONMENT_CONTEXT);
     }
 
-    
+
     @Override
     @JsonIgnore
     public boolean isValid() {
