@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.util.Map;
 
 import org.openape.server.rest.SuperRestInterface;
+import org.openape.ui.velocity.OpenApeVelocityEngine;
 import org.openape.ui.velocity.controllerComponents.MainComponents;
 
 import spark.ModelAndView;
@@ -25,7 +26,7 @@ public class Downloads extends SuperRestInterface {
                                                                               // resources
                                                                               // directory
 
-            }, new VelocityTemplateEngine());
+            }, new OpenApeVelocityEngine());
     }
 
     public Downloads() throws IllegalArgumentException, IOException {
