@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.util.Map;
 
 import org.openape.server.rest.SuperRestInterface;
+import org.openape.ui.velocity.OpenApeVelocityEngine;
 import org.openape.ui.velocity.controllerComponents.MainComponents;
 
 import spark.ModelAndView;
@@ -13,7 +14,7 @@ import spark.template.velocity.VelocityTemplateEngine;
 public class Index extends SuperRestInterface {
     public static void setupIndexVELOCITYInterface() throws IllegalArgumentException, IOException {
 
-        Spark.get("/index", (request, response) -> {
+        Spark.get("/", (request, response) -> {
             final MainComponents mainController = new MainComponents();
             final Map<String, Object> model = mainController.getTemplateComponents();
 
@@ -23,7 +24,7 @@ public class Index extends SuperRestInterface {
                                                                           // resources
                                                                           // directory
 
-            }, new VelocityTemplateEngine());
+            }, new OpenApeVelocityEngine());
     }
 
     public Index() throws IllegalArgumentException, IOException {
