@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.util.Map;
 
 import org.openape.server.rest.SuperRestInterface;
+import org.openape.ui.velocity.OpenApeVelocityEngine;
 import org.openape.ui.velocity.controllerComponents.MainComponents;
 import org.openape.ui.velocity.organism.Organism_3_DataTable;
 import org.openape.ui.velocity.organism.Organism_4_Modals;
@@ -41,7 +42,7 @@ public class MyResources extends SuperRestInterface {
                                                                                         // the
                                                                                         // resources
                                                                                         // directory
-                }, new VelocityTemplateEngine());
+                }, new OpenApeVelocityEngine());
     }
 
     public MyResources() throws IllegalArgumentException, IOException {
