@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.util.Map;
 
 import org.openape.server.rest.SuperRestInterface;
+import org.openape.ui.velocity.OpenApeVelocityEngine;
 import org.openape.ui.velocity.controllerComponents.MainComponents;
 
 import spark.ModelAndView;
@@ -14,11 +15,11 @@ public class Tutorials_Workflow extends SuperRestInterface {
     public static void setupTutorialsWorkflowVELOCITYInterface() throws IllegalArgumentException,
             IOException {
 
-        Spark.get("/workflow", (request, response) -> { return getTutorial("workflow.vm");},new VelocityTemplateEngine());
-        Spark.get("/openape_js", (request, response) -> { return getTutorial("openape_js.vm");},new VelocityTemplateEngine());
-        Spark.get("/userContextTutorial", (request, response) -> { return getTutorial("userContextTutorial.vm");},new VelocityTemplateEngine());
+        Spark.get("/workflow", (request, response) -> { return getTutorial("workflow.vm");},new OpenApeVelocityEngine());
+        Spark.get("/openape_js", (request, response) -> { return getTutorial("openape_js.vm");},new OpenApeVelocityEngine());
+        Spark.get("/userContextTutorial", (request, response) -> { return getTutorial("userContextTutorial.vm");},new OpenApeVelocityEngine());
     }
-    
+
     static ModelAndView getTutorial(String fileName) {
         final Map<String, Object> model = new MainComponents().getTemplateComponents();
 
